python-backend/src/requirements.txt
supabase
fastapi
uvicorn[standard]
boto3
pydantic
human-id
pydantic_settings
requests
google-cloud-artifact-registry==1.8.3